requirements.txt
yfinance>=0.2.28
pandas>=2.0.0
schedule>=1.2.0
colorama>=0.4.6
python-dateutil>=2.8.2
pytz>=2023.3
pyyaml>=6.0
